requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
plotly>=5.13.0
numpy>=1.21.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
torch>=2.0.0